requirements.txt
certifi==2023.7.22
click==8.1.7
Flask==3.0.0
idna==3.4
ImageHash==4.3.1
imageio==2.31.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
networkx==3.1
Pillow==10.0.0
pymongo==4.5.0
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.2
six==1.16.0
toolz==0.12.0
urllib3==1.26.18
Werkzeug==3.0.1
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
apyori==1.1.2
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bitarray==2.8.1
blinker==1.6.2
blis==0.7.10
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.7
confection==0.1.3
contourpy==1.1.1
cssselect==1.2.0
cycler==0.12.1
cymem==2.0.8
cytoolz==0.12.2
distlib==0.3.7
Django==4.2.6
dnspython==2.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
et-xmlfile==1.1.0
eth-abi==4.2.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
filelock==3.12.3
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.5.4
GetOldTweets3==0.0.11
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.10.0
hexbytes==0.3.1
hmmlearn==0.3.0
huggingface-hub==0.17.3
idna==3.4
ImageHash==4.3.1
imageio==2.31.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.14.0
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
lru-dict==1.2.0
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
multidict==6.0.4
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
panda==0.3.1
pandas==2.1.1
parsimonious==0.9.0
pathy==0.10.2
Pillow==10.0.0
pipenv==2023.10.3
platformdirs==3.10.0
preshed==3.0.9
protobuf==4.24.2
pyacoustid==1.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
pyparsing==3.1.1
pypng==0.20220715.0
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyunormalize==15.0.0
PyWavelets==1.4.1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rlp==3.0.0
rpds-py==0.10.0
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.13.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.20
sqlparse==0.4.4
srsly==2.4.7
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
textblob==0.17.1
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.14.1
toolz==0.12.0
tqdm==4.66.1
transformers==4.34.0
tweepy==4.14.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.18
virtualenv==20.24.4
wasabi==1.1.2
web3==6.9.0
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.14.1
WTForms==3.0.1
yarl==1.9.2
